# api.yaml
openapi: 3.0.3
info:
  title: Go Auth Service API
  description: |
    API documentation for the Go Auth Service - a production-ready Identity Provider.
    
    ## Features
    - Authentication (Register, Login, Refresh, Logout)
    - Email Verification
    - Multi-tenant Management
    - RBAC (Role-Based Access Control)
  version: 1.0.0
  contact:
    name: Michael
    url: https://t.me/d1ad3m

servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # --- Common ---
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid credentials"
        code:
          type: string
          example: "AUTH_INVALID_CREDENTIALS"
        details:
          type: object
          additionalProperties: true

    # --- Auth ---
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        is_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    
    TokenPair:
      type: object
      required:
        - access_token
        - refresh_token
        - expires_in
        - token_type
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Access token expiration in seconds
        token_type:
          type: string
          example: "Bearer"

    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string

    VerifyEmailRequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          description: Verification code sent to email
    
    ResendVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email

    # --- Tenants ---
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
    
    CreateTenantRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          minLength: 3
        slug:
          type: string
          minLength: 3
          pattern: "^[a-z0-9-]+$"

paths:
  # --- System ---
  /health:
    get:
      summary: Health Check
      description: Check if the service is up and running
      tags:
        - System
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"

  # --- Authentication ---
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  message:
                    type: string
                    example: "Please check your email to verify your account"
        '400':
          description: Bad request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      summary: Login user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/refresh:
    post:
      summary: Refresh access token
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      summary: Logout user
      security:
        - BearerAuth: []
      tags:
        - Auth
      responses:
        '204':
          description: Logged out successfully
        '401':
          description: Unauthorized

  /auth/verify-email:
    post:
      summary: Verify email address
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified successfully
        '400':
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/resend-verification:
    post:
      summary: Resend verification email
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
      responses:
        '200':
          description: Verification email sent (if user exists and not verified)
        '429':
          description: Too many requests

  # --- Users ---
  /users/me:
    get:
      summary: Get current user profile
      security:
        - BearerAuth: []
      tags:
        - Users
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

  # --- Tenants ---
  /tenants:
    get:
      summary: List user's tenants
      security:
        - BearerAuth: []
      tags:
        - Tenants
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tenant'
        '401':
          description: Unauthorized
    
    post:
      summary: Create a new tenant
      security:
        - BearerAuth: []
      tags:
        - Tenants
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized