# internal/domain/tenant.go
package domain

import (
	"time"
)

// Tenant represents an organization or workspace.
type Tenant struct {
	ID        string    `json:"id"`
	Name      string    `json:"name"`
	Slug      string    `json:"slug"`
	CreatedAt time.Time `json:"created_at"`
}

// NewTenant creates a new tenant.
func NewTenant(name, slug string) *Tenant {
	return &Tenant{
		// ID should be generated by the caller
		Name:      name,
		Slug:      slug,
		CreatedAt: time.Now().UTC(),
	}
}