# internal/domain/user.go
package domain

import (
	"time"
)

// User represents a registered user in the system.
type User struct {
	ID         string    `json:"id"`
	Email      string    `json:"email"`
	Password   string    `json:"-"` // Never return password hash
	IsVerified bool      `json:"is_verified"`
	CreatedAt  time.Time `json:"created_at"`
	UpdatedAt  time.Time `json:"updated_at"`
}

// NewUser creates a new user instance with default values.
func NewUser(email, passwordHash string) *User {
	now := time.Now().UTC()
	return &User{
		// ID should be generated by the caller or repository (UUID)
		Email:      email,
		Password:   passwordHash,
		IsVerified: false,
		CreatedAt:  now,
		UpdatedAt:  now,
	}
}